requirements.txt
python-dotenv
openai
firecrawl-py
beautifulsoup4
langchain
langchain-community
langchain-openai
chromadb
requests
flask
flask-cors
sseclient-py
gunicorn
pytest
pytest-mock
pytest-asyncio
pytest-cov
gevent
redis
aioredis